Updated TimeMethod, yet again.
TimeMethod is now a functional interface because time() is no longer a method. I realized that there's actually no need to have that method since, as far as I know, all timing units are just scalars of each other. Internally, now everythign is timed in nanoseconds and then converted using the convert method.