Unit Tests
Added a rather large amount of unit tests. The unit tests for the graph use a Robot instead of MouseEvents, because I think that's a better way to test it. I think JTimer is currently at a good place now so I wouldn't be updating it quite as often, althuogh updates will still be going.